api/pyproject.toml
[project]
name = "lecturia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "neka-nat", email = "nekanat.stock@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "google-genai>=1.7.0",
    "langchain>=0.3.21",
    "langchain-anthropic>=0.3.10",
    "loguru>=0.7.3",
    "moviepy>=1.0.3",
    "numpy>=2.2.4",
    "openai>=1.68.0",
    "pillow>=11.1.0",
    "playwright>=1.51.0",
    "pydantic>=2.10.6",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
